pyproject.toml
[project]
name = "multimodal-rag-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anyio>=4.13.0",
    "google-generativeai>=0.8.6",
    "google-cloud-aiplatform>=1.93.0",
    "openai>=2.30.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "google-genai>=1.69.0",
    "qdrant-client>=1.17.1",
    "elasticsearch>=8.0.0,<9.0.0",
    "ranx>=0.3.21",
    "evaluate>=0.4.6",
    "ragas>=0.4.3",
    "pymupdf>=1.24.0",
    "docling>=2.84.0",
    "matplotlib>=3.10.8",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "boto3>=1.42.89",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pillow>=10.2.0.20240822",
    "types-pyyaml>=6.0.12.20250915",
]
